chore: attempt at separating factorizations
